control-service: move cron jobs methods to the data jobs class Signed-off-by: murphp15 <murphp15@tcd.ie>
